% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_yymmdd.R
\name{date_to_yymmdd}
\alias{date_to_yymmdd}
\title{Converts a date in yyyy-mm-dd format to yymmdd character string}
\usage{
date_to_yymmdd(date)
}
\arguments{
\item{date}{date to convert}
}
\value{
a date in yymmdd character format.
}
\description{
Converts a date in yyyy-mm-dd format to yymmdd character string
}
\examples{


}
