% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSTdate_to_date.R
\name{DSTdate_to_date}
\alias{DSTdate_to_date}
\title{Converts a date in typical Danmarks Statistik format to yyyy-mm-dd date format}
\usage{
DSTdate_to_date(DSTdate)
}
\arguments{
\item{DSTdate}{character. Date of interest.}
}
\value{
a date in yyyy-mm-dd date format.
}
\description{
Converts a date in typical Danmarks Statistik format to yyyy-mm-dd date format
}
\examples{


}
