% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor_date_wday.R
\name{floor_date_wday}
\alias{floor_date_wday}
\title{Converts a date to the date of the last give weekday}
\usage{
floor_date_wday(date, day = 1)
}
\arguments{
\item{date}{numeric. Weekday of interest.}
}
\value{
a date in yyyy-mm-dd date format.
}
\description{
Converts a date to the date of the last give weekday
}
\examples{


}
