% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_breaks.R
\name{get_age_breaks}
\alias{get_age_breaks}
\title{Helper function for get_pop_by_breaks().}
\usage{
get_age_breaks(maxage = 100, agesplit = 10)
}
\arguments{
\item{maxage}{numeric. Lower limit of maximum age bracket.}

\item{agesplit}{numeric. Window size of age brackets.}
}
\value{
a numeric vector of age breaks
}
\description{
Helper function for get_pop_by_breaks().
}
\examples{


}
