% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_by_breaks.R
\name{get_pop_by_breaks}
\alias{get_pop_by_breaks}
\title{Function that takes population data from Danmarks Statistik and calculates population sizes in given brackets}
\usage{
get_pop_by_breaks(data = "../data/tidy_DST_pop.csv", age_breaks)
}
\arguments{
\item{data}{dataframe. Tidy dataframe of DST population data}

\item{age_breaks}{numeric vector. Upper limit of each age bracket}
}
\value{
a dataframe with population sizes by age brackets
}
\description{
Function that takes population data from Danmarks Statistik and calculates population sizes in given brackets
}
\examples{


}
