% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sample.R
\name{mc_sample}
\alias{mc_sample}
\title{Sample}
\usage{
mc_sample(
  nrep,
  input,
  pop,
  reference_date,
  first_date,
  end_prediction_date,
  contact_reduction_first_day,
  rnd_func = crunif
)
}
\arguments{
\item{nrep}{Antal bootstraps}

\item{input}{Parameters}
}
\description{
Sample
}
