% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_date_labels_en.R
\name{my_date_labels_en}
\alias{my_date_labels_en}
\title{Function to add date labels on ggplots in English with the year added on the first label of a new year}
\usage{
my_date_labels_en(breaks)
}
\arguments{
\item{breaks}{breaks as given by the breaks argument in scale_x_date}
}
\value{
a vector of labels
}
\description{
Function to add date labels on ggplots in English with the year added on the first label of a new year
}
\examples{

}
