% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_date_labels_no_year.R
\name{my_date_labels_no_year}
\alias{my_date_labels_no_year}
\title{Function to add date labels on ggplots in Danish without year}
\usage{
my_date_labels_no_year(breaks)
}
\arguments{
\item{breaks}{breaks as given by the breaks argument in scale_x_date}
}
\value{
a vector of labels
}
\description{
Function to add date labels on ggplots in Danish without year
}
\examples{

}
