% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra.R
\name{ra}
\alias{ra}
\title{Calculate running average}
\usage{
ra(x, n = 7, s = 2, ...)
}
\arguments{
\item{x}{vector to calculate the unning average on.}

\item{n}{window size. Default = 7.}

\item{s}{side. If s = 1, the values are based on past values only, if s = 2, the values are centered in the window.}

\item{...}{further arguments passed to stats::filter}
}
\value{
a vector of running average values
}
\description{
Calculate running average
}
\examples{


}
