% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Reads csv files}
\usage{
read_data(path, type = 2, ...)
}
\arguments{
\item{path}{path to file}

\item{type}{numeric. 1 uses read_csv, 2 uses read_csv2.}
}
\value{
a dataframe.
}
\description{
Reads csv files
}
\examples{


}
