% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_echo.R
\name{source_echo}
\alias{source_echo}
\title{Source a file and print filepath}
\usage{
source_echo(filepath)
}
\arguments{
\item{filepath}{path to R file}
}
\value{

}
\description{
Source a file and print filepath
}
\examples{


}
