% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_to_date.R
\name{week_to_date}
\alias{week_to_date}
\title{Converts a week number to a date in that week}
\usage{
week_to_date(year, week, day = "1")
}
\arguments{
\item{year}{numeric or character. Year of interest.}

\item{week}{numeric or character. Week of interest.}

\item{day}{numeric or character. Day of week to get date of. Monday = 1 etc.}
}
\value{
a date in yyyy-mm-dd date format.
}
\description{
Converts a week number to a date in that week
}
\examples{


}
